# backend/index.js
const express = require('express');
const cors = require('cors');
const multer = require('multer');
const fs = require('fs');
const path = require('path');
const jobManager = require('./modules/jobs/jobManager');
const { processJobFiles } = require('./modules/fileProcessor');
const logger = require('./modules/logger');

const app = express();
const PORT = 4000;

app.use(cors());
app.use(express.json());

const uploadDir = path.join(__dirname, 'uploads');
if (!fs.existsSync(uploadDir)) {
    fs.mkdirSync(uploadDir, { recursive: true });
}

const storage = multer.diskStorage({
    destination: function (req, file, cb) {
        cb(null, uploadDir)
    },
    filename: function (req, file, cb) {
        cb(null, Date.now() + '-' + file.originalname)
    }
});
const upload = multer({
    storage: storage,
    limits: { fileSize: 50 * 1024 * 1024 }
});

app.post('/api/v1/jobs', upload.array('files'), async (req, res) => {
    try {
        if (!req.files || req.files.length === 0) {
            logger.warn('Upload attempted with no files');
            return res.status(400).json({ error: 'No files provided' });
        }

        logger.info('Files uploaded', { count: req.files.length, filenames: req.files.map(f => f.originalname) });

        const files = req.files.map(file => ({
            originalName: file.originalname,
            uploadPath: file.path,
            uploadedAt: new Date().toISOString(),
        }));

        const job = jobManager.createJob(files);
        logger.info('Job created', { jobId: job.id, fileCount: files.length });

        const filePaths = req.files.map(file => ({
            filePath: file.path,
            originalFilename: file.originalname,
        }));

        processJobFiles(job.jobId, filePaths).catch(err => {
            console.error('Error processing job:', err);
            jobManager.updateJobStatus(job.jobId, jobManager.JOB_STATUSES.FAILED);
        });

        res.status(201).json({
            jobId: job.jobId,
            status: job.status,
            totalFiles: job.files.length,
            message: 'Job created. Processing started...',
        });
    } catch (err) {
        logger.error('Upload error', { error: err.message, stack: err.stack });
        res.status(500).json({ error: 'Failed to create job', details: err.message });
    }
});

app.get('/api/v1/jobs/:id', (req, res) => {
    try {
        const jobId = req.params.id;
        const summary = jobManager.getJobSummary(jobId);

        if (!summary) {
            logger.warn('Job not found', { jobId });
            return res.status(404).json({ error: 'Job not found' });
        }

        logger.debug('Job summary fetched', { jobId, status: summary.status });
        res.json(summary);
    } catch (err) {
        logger.error('Error fetching job', { jobId: req.params.id, error: err.message });
        res.status(500).json({ error: 'Failed to fetch job' });
    }
});

app.get('/api/v1/jobs/:id/export', (req, res) => {
    try {
        const jobId = req.params.id;
        const job = jobManager.loadJob(jobId);

        if (!job) {
            return res.status(404).json({ error: 'Job not found' });
        }

        if (job.status !== jobManager.JOB_STATUSES.COMPLETED) {
            return res.status(409).json({ error: 'Job is not completed yet', status: job.status, code: 'JOB_NOT_COMPLETED' });
        }

        const exportData = jobManager.exportJobAsJson(jobId);
        res.json(exportData);
    } catch (err) {
        console.error('Error exporting job:', err);
        res.status(500).json({ error: 'Failed to export job' });
    }
});

app.put('/api/v1/jobs/:id/annotation', express.json(), (req, res) => {
    try {
        const jobId = req.params.id;
        const { fileIndex, annotation } = req.body;

        if (fileIndex === undefined || !annotation) {
            return res.status(400).json({ error: 'File index and annotation required' });
        }

        const updated = jobManager.updateFileAnnotation(jobId, fileIndex, annotation);
        if (!updated) {
            return res.status(404).json({ error: 'Job or file not found' });
        }

        res.json({ success: true, message: 'Annotation saved' });
    } catch (err) {
        console.error('Error saving annotation:', err);
        res.status(500).json({ error: 'Failed to save annotation' });
    }
});

app.get('/api/v1/jobs', (req, res) => {
    try {
        const limit = req.query.limit ? parseInt(req.query.limit) : 50;
        const jobs = jobManager.listJobs(limit);
        res.json({ jobs, total: jobs.length });
    } catch (err) {
        console.error('Error listing jobs:', err);
        res.status(500).json({ error: 'Failed to list jobs' });
    }
});

app.get('/api/v1/health', (req, res) => {
    res.json({
        status: 'ok',
        server: 'AINEVERCRY Gateway',
        version: '1.0.0',
        uptime: process.uptime(),
    });
});

app.get('/docs', (req, res) => {
    const html = `
    <!DOCTYPE html>
    <html>
      <head>
        <title>Data Refinery API Docs</title>
        <meta charset="utf-8"/>
        <meta name="viewport" content="width=device-width, initial-scale=1">
        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swagger-ui-dist@3/swagger-ui.css">
        <style>
          body {
            margin: 0;
            padding: 0;
          }
        </style>
      </head>
      <body>
        <div id="swagger-ui"></div>
        <script src="https://cdn.jsdelivr.net/npm/swagger-ui-dist@3/swagger-ui-bundle.js"></script>
        <script>
          window.onload = function() {
            SwaggerUIBundle({
              url: '/openapi.yaml',
              dom_id: '#swagger-ui',
              presets: [
                SwaggerUIBundle.presets.apis,
                SwaggerUIBundle.SwaggerUIStandalonePreset
              ],
              layout: "BaseLayout",
              deepLinking: true
            })
          }
        </script>
      </body>
    </html>
    `;
    res.setHeader('Content-Type', 'text/html');
    res.send(html);
});

app.get('/openapi.yaml', (req, res) => {
    const spec = fs.readFileSync(path.join(__dirname, '../openapi.yaml'), 'utf-8');
    res.setHeader('Content-Type', 'application/yaml');
    res.send(spec);
});

app.listen(PORT, () => {
    logger.info(`Server started on http://localhost:${PORT}`);
    console.log(`\n🚀 AINEVERCRY Backend started on http://localhost:${PORT}`);
    console.log(`📍 API Base: http://localhost:${PORT}/api/v1`);
    console.log(`\n📚 Available endpoints:`);
    console.log(`  POST   /api/v1/jobs              - Create job with files`);
    console.log(`  GET    /api/v1/jobs/:id          - Get job status`);
    console.log(`  GET    /api/v1/jobs/:id/export   - Export final JSON`);
    console.log(`  GET    /api/v1/jobs              - List all jobs`);
    console.log(`  GET    /api/v1/health            - Health check\n`);
});