# openapi.yaml
openapi: 3.0.3
info:
  title: Data Refinery Gateway API
  description: Secure data processing gateway for document and image sanitization
  version: 1.0.0
  contact:
    name: AINEVERCRY
  license:
    name: MIT

servers:
  - url: http://localhost:4000/api/v1
    description: Local development

tags:
  - name: Jobs
    description: Job lifecycle management and processing
  - name: Health
    description: Service status monitoring

paths:
  /jobs:
    post:
      tags:
        - Jobs
      summary: Create processing job
      description: |
        Initiates a new data processing job with uploaded files.
        Files are queued for sequential processing immediately after job creation.
        Processing includes validation, content extraction, PII detection, metadata removal, and image normalization.
      operationId: createJob
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to process (PDF, TXT, MD, CSV, JPG, PNG, SVG). Max 50MB each.
                  minItems: 1
                  maxItems: 100
              required:
                - files
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Jobs
      summary: List all jobs
      description: Retrieves paginated list of all processing jobs with status
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Jobs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCount:
                    type: integer
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobSummary'

  /jobs/{jobId}:
    get:
      tags:
        - Jobs
      summary: Get job status
      description: |
        Retrieves current job status and per-file processing progress.
        Poll endpoint every 1-2 seconds until status is 'completed' or 'failed'.
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /jobs/{jobId}/export:
    get:
      tags:
        - Jobs
      summary: Export processed data
      description: |
        Downloads unified JSON containing all processed data transformations.
        Available only when job status is 'completed'.
      operationId: exportJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Processed data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job not completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /jobs/{jobId}/annotation:
    put:
      tags:
        - Jobs
      summary: Update file annotation
      description: |
        Add or update annotation for a specific file within a job.
        Annotations are operator notes stored with the job and exported in final JSON.
      operationId: updateAnnotation
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fileIndex
                - annotation
              properties:
                fileIndex:
                  type: integer
                  minimum: 0
                  description: Zero-based index of file in job
                annotation:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: Annotation text
      responses:
        '200':
          description: Annotation saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job or file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns service status and version info
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok, degraded]
                  version:
                    type: string
                  uptime:
                    type: number
                    description: Uptime in seconds

components:
  schemas:
    JobCreatedResponse:
      type: object
      required:
        - jobId
        - status
        - totalFiles
      properties:
        jobId:
          type: string
          format: uuid
          description: Unique identifier for the created job
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Initial job status
        totalFiles:
          type: integer
          minimum: 1
          description: Number of files uploaded
        message:
          type: string
          description: Status message
        files:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              originalName:
                type: string
              status:
                type: string
                enum: [pending, processing, processed, failed]
          nullable: true

    JobStatusResponse:
      type: object
      required:
        - jobId
        - status
        - totalFiles
        - processedFiles
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalFiles:
          type: integer
        processedFiles:
          type: integer
        failedFiles:
          type: integer
        progress:
          type: integer
          minimum: 0
          maximum: 100
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileStatus'

    FileStatus:
      type: object
      required:
        - id
        - originalName
        - status
        - hasErrors
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        detectedType:
          type: string
          enum: [PDF, TXT, MD, CSV, JPG, PNG, SVG]
          nullable: true
        status:
          type: string
          enum: [pending, processing, processed, failed]
        hasErrors:
          type: boolean

    ExportResponse:
      type: object
      required:
        - system
        - version
        - jobId
        - processedAt
        - totalFiles
        - files
        - summary
      properties:
        system:
          type: string
          example: Data Refinery Gateway
        version:
          type: string
          example: 1.0.0
        jobId:
          type: string
          format: uuid
        processedAt:
          type: string
          format: date-time
        totalFiles:
          type: integer
        files:
          type: array
          items:
            $ref: '#/components/schemas/ExportedFile'
        summary:
          $ref: '#/components/schemas/ProcessingSummary'

    ExportedFile:
      type: object
      required:
        - id
        - originalName
        - type
        - secure
        - processing
      properties:
        id:
          type: string
          format: uuid
        originalName:
          type: string
        type:
          type: string
          enum: [PDF, TXT, MD, CSV, IMAGE, SVG]
        mimeType:
          type: string
        secure:
          type: boolean
        processing:
          type: object
          properties:
            validated:
              type: boolean
            extracted:
              type: boolean
            sanitized:
              type: boolean
            piiRedacted:
              type: boolean
        metadata:
          type: object
          nullable: true
          properties:
            annotation:
              type: string
            fileSize:
              type: integer
            processedAt:
              type: string
              format: date-time
        extracted:
          type: object
          nullable: true
        warnings:
          type: array
          items:
            type: string
          nullable: true
        validationErrors:
          type: array
          items:
            type: string
          nullable: true
        securityIssues:
          type: array
          items:
            $ref: '#/components/schemas/SecurityIssue'
          nullable: true

    SecurityIssue:
      type: object
      required:
        - type
        - message
      properties:
        type:
          type: string
          enum: [XSS, SQL_INJECTION, POLYGLOT, EXECUTABLE, DOUBLE_EXTENSION]
        message:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
          default: MEDIUM

    ProcessingSummary:
      type: object
      required:
        - totalProcessed
        - totalFailed
        - secureFiles
      properties:
        totalProcessed:
          type: integer
        totalFailed:
          type: integer
        secureFiles:
          type: integer
        filesWithPII:
          type: integer
        filesWithSecurityIssues:
          type: integer
        filesWithMetadata:
          type: integer

    JobSummary:
      type: object
      required:
        - jobId
        - status
        - createdAt
        - totalFiles
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        totalFiles:
          type: integer
        processedFiles:
          type: integer
        progress:
          type: integer

    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
            - JOB_NOT_FOUND
            - JOB_NOT_COMPLETED
            - VALIDATION_FAILED
            - SECURITY_BLOCKED
            - PROCESSING_ERROR
            - INTERNAL_ERROR
        details:
          type: object
          nullable: true
        status:
          type: string
          nullable: true